AltairZ80: Addressed Microsoft compiler generated warnings

+++ AltairZ80/m68kopac.c
@@ -5691,14 +5691,14 @@ void m68k_op_bfexts_32_d(void)
         width = ((width-1) & 31) + 1;
 
         data = ROL_32(data, offset);
-        FLAG_N = NFLAG_32(data);
+        FLAG_N = (uint)(NFLAG_32(data));
         data = MAKE_INT_32(data) >> (32 - width);
 
-        FLAG_Z = data;
+        FLAG_Z = (uint)data;
         FLAG_V = VFLAG_CLEAR;
         FLAG_C = CFLAG_CLEAR;
 
-        REG_D[(word2>>12)&7] = data;
+        REG_D[(word2>>12)&7] = (uint)data;
 
         return;
     }
@@ -6061,14 +6061,14 @@ void m68k_op_bfextu_32_d(void)
         width = ((width-1) & 31) + 1;
 
         data = ROL_32(data, offset);
-        FLAG_N = NFLAG_32(data);
+        FLAG_N = (uint)(NFLAG_32(data));
         data >>= 32 - width;
 
-        FLAG_Z = data;
+        FLAG_Z = (uint)data;
         FLAG_V = VFLAG_CLEAR;
         FLAG_C = CFLAG_CLEAR;
 
-        REG_D[(word2>>12)&7] = data;
+        REG_D[(word2>>12)&7] = (uint)data;
 
         return;
     }
@@ -6425,10 +6425,10 @@ void m68k_op_bfffo_32_d(void)
         width = ((width-1) & 31) + 1;
 
         data = ROL_32(data, offset);
-        FLAG_N = NFLAG_32(data);
+		FLAG_N = (uint)(NFLAG_32(data));
         data >>= 32 - width;
 
-        FLAG_Z = data;
+		FLAG_Z = (uint)data;
         FLAG_V = VFLAG_CLEAR;
         FLAG_C = CFLAG_CLEAR;
 
@@ -6833,8 +6833,8 @@ void m68k_op_bfins_32_d(void)
         mask = ROR_32(mask, offset);
 
         insert = MASK_OUT_ABOVE_32(insert << (32 - width));
-        FLAG_N = NFLAG_32(insert);
-        FLAG_Z = insert;
+		FLAG_N = (uint)(NFLAG_32(insert));
+		FLAG_Z = (uint)insert;
         insert = ROR_32(insert, offset);
 
         FLAG_V = VFLAG_CLEAR;
